+++ frontend/src/pages/projects/screens/detail/StorageList.tsx
@@ -4,14 +4,20 @@ import EmptyDetailsList from './EmptyDetailsList';
 import DetailsSection from './DetailsSection';
 import { ProjectSectionID } from './types';
 import { ProjectSectionTitles } from './const';
+import AddStorageModal from 'pages/projects/modals/addStorageModal/AddStorageModal';
 
 const StorageList: React.FC = () => {
+  const [isOpen, setOpen] = React.useState<boolean>(false);
   return (
     <DetailsSection
       id={ProjectSectionID.STORAGE}
       title={ProjectSectionTitles[ProjectSectionID.STORAGE]}
       actions={[
-        <Button key={`action-${ProjectSectionID.STORAGE}`} variant="secondary">
+        <Button
+          onClick={() => setOpen(true)}
+          key={`action-${ProjectSectionID.STORAGE}`}
+          variant="secondary"
+        >
           Add storage
         </Button>,
       ]}
@@ -20,6 +26,7 @@ const StorageList: React.FC = () => {
         title="No storage"
         description="Choose existing, or add new on cluster storage."
       />
+      <AddStorageModal isOpen={isOpen} onClose={() => setOpen(false)} />
     </DetailsSection>
   );
 };
